#1396 Moved extraction of Sound object to a dedicated class
